# docs/openapi/tracebud-v1-draft.yaml
openapi: 3.0.3
info:
  title: Tracebud API (v1 Draft)
  version: "0.1.0-draft"
  description: |
    Draft OpenAPI artifact generated from TRACEBUD_V1_2_EUDR_SPEC.md Section 32.8.
    This document is intentionally minimal and should be expanded during implementation.
  license:
    name: Proprietary
    url: https://tracebud.com/legal
  x-tracebud-error-mapping:
    AUTH-001: INVALID_TOKEN
    AUTH-002: INSUFFICIENT_SCOPE
    VAL-001: FIELD_REQUIRED
    VAL-002: FIELD_INVALID
    TEN-001: TENANT_SCOPE_VIOLATION
    CON-001: CONSENT_REQUIRED
    SYNC-001: IDEMPOTENCY_REPLAY
    SYNC-002: CONFLICT_REVIEW_REQUIRED
    GEO-001: GEOMETRY_INVALID
    GEO-002: GEOMETRY_VARIANCE_EXCEEDED
    YLD-001: BENCHMARK_UNAVAILABLE
    DDS-001: INVALID_STATE_TRANSITION
    DDS-002: SUBMISSION_REJECTED
    EXT-001: TRACES_AUTH_FAILED
    EXT-002: TRACES_VALIDATION_FAULT
    BIL-001: PAYMENT_FAILED
    OPS-001: INTERNAL_ERROR
servers:
  - url: https://api.tracebud.com
    description: Production API
  - url: https://sandbox-api.tracebud.com
    description: Sandbox API
security:
  - bearerAuth: []
...
paths:
...
  /v1/integrations/assessments/requests:
    post:
      tags: [Integrations]
      operationId: createAssessmentRequest
      summary: Create assessment request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - farmerId
                - pathway
              properties:
                title:
                  type: string
                instructions:
                  type: string
                farmerId:
                  type: string
                pathway:
                  type: string
                  enum: [annuals, rice]
                dueDate:
                  type: string
                  format: date
                questionnaireDraftId:
                  type: string
      responses:
        "201":
          description: Assessment request created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Missing tenant claim or role policy violation
    get:
      tags: [Integrations]
      operationId: listAssessmentRequests
      summary: List assessment requests
      parameters:
        - name: assignedToMe
          in: query
          required: false
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Tenant-scoped assessment requests
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Missing tenant claim or role policy violation
...
  /v1/integrations/assessments/requests/{id}/status:
    patch:
      tags: [Integrations]
      operationId: updateAssessmentRequestStatus
      summary: Update assessment request status (manager)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [sent, opened, in_progress, submitted, reviewed, needs_changes, cancelled]
      responses:
        "200":
          description: Assessment status updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Missing tenant claim or role policy violation
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error_code: { type: string }
        message: { type: string }
        field_errors:
          type: array
          items: { type: object }
        request_id: { type: string }
      required: [error_code, message, request_id]
...